+++ ReleaseNotes.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+## 0.00.002
+
+* add start and runntime info to tasklist
+* add failed state (exclamation mark)
+* add colors to state icons
+
 ## 0.00.001
 
 * basic working app